# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovelOS API",
    "version": "3.8.0"
  },
  "paths": {
...
    "/api/model-configs": {
      "post": {
        "tags": [
          "model_configs"
        ],
        "summary": "Create Model Config",
        "description": "创建一条 model config。必填字段：``capability / provider / model``。\n\n入参 ``params_json.api_key`` 的特殊值：\n- ``\"\"`` 或 ``\"***\"``（:data:`_MASK`）→ 视为未设置，落库时不写入 api_key。\n- 其他字符串 → 原样落库（明文）。",
        "operationId": "create_model_config_api_model_configs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Model Config Api Model Configs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "model_configs"
        ],
        "summary": "List Model Configs",
        "operationId": "list_model_configs_api_model_configs_get",
        "parameters": [
          {
            "name": "capability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capability"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Model Configs Api Model Configs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/model-configs/{config_id}": {
      "get": {
        "tags": [
          "model_configs"
        ],
        "summary": "Get Model Config",
        "operationId": "get_model_config_api_model_configs__config_id__get",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Model Config Api Model Configs  Config Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "model_configs"
        ],
        "summary": "Patch Model Config",
        "description": "部分更新。\n\n入参 ``params_json.api_key`` 的特殊值：\n- ``\"***\"``（:data:`_MASK`）→ 视为「保留 DB 原值」，不动 api_key；\n- 空字符串 → 视为「清空 api_key」（删除该键）；\n- 其他 → 用入参值覆盖。\n\n其余字段（capability / provider / model / enabled）沿用既有非空校验。",
        "operationId": "patch_model_config_api_model_configs__config_id__patch",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Model Config Api Model Configs  Config Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "model_configs"
        ],
        "summary": "Delete Model Config",
        "operationId": "delete_model_config_api_model_configs__config_id__delete",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/model-configs/{config_id}/test": {
      "post": {
        "tags": [
          "model_configs"
        ],
        "summary": "Test Model Config",
        "description": "ping：用该配置做一次轻量级健康检查，返回 ``{ok, latency_ms, detail, status_code?}``。\n\n行为契约（Sprint 8）：\n- ``mock`` provider → 直接 ``ok=True``（不发起真实调用，与既有 mock 语义一致）。\n- 其它 provider → 调 :meth:`Provider.health_check`：OpenAI 兼容 ``GET /models``、\n  Ollama ``GET /api/tags``、Anthropic 极小 ``POST /v1/messages``（401/200 都算通）。\n- 健康检查失败 → 200 但 ``ok=False``（前端可区分「端点不可达」与「鉴权失败」）；\n  仅 ProviderError 透传到 502 兜底保留。\n- ``enabled=0`` 仍返回 422（P2-5 业务规则）。",
        "operationId": "test_model_config_api_model_configs__config_id__test_post",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Test Model Config Api Model Configs  Config Id  Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}